libproc proxying, bringing ustack() support.
This change implements a simple proxying mechanism in dt_proc, using the
interfaces added in commit b924c3755115dc5b204f74b65565cf80dba4939b.  The core
proxying mechanism consists of ptrace() and Pwait() wrappers which work in very
similar ways.  They descibe their parameters in the new tagged union
dt_proc_t.dpr_proxy_args, and indicate which hook they are by assigning their
own address to dt_proc_t.dpr_proxy_rq: thus, the name of the function (thus, its
address) can be used as an indication of what the proxied operation was.  Then
they spit a single byte down the new dpr_proxy_fd pipe and go to sleep on the
new dpr_msg_cv condition variable. When awoken, the proxy operation is done
and all they have to do is demarshal the return value and errno, and return.
(Because we use only one return value, we must adjust Pwait() so that it returns
a long, rather than an int, just like ptrace().  We may be forced to separate
the return values later, if we add more proxies, but for now this was simpler.)

When called when the current thread is the control thread, they just forward to
the internal functions, so the same caller can operate without considering
whether it is in the control thread or not, and just call the proxy and be sure
it will do the right thing.

The control thread changes from something which waits in Pwait() forever to
something which poll()s on two file descriptors, one being the other end of the
dpr_proxy_fd pipe and the other being a waitfd() on the process this thread is
controlling.  When a request comes in, we first check to see if it is a proxy
request before consuming a process state change using Pwait() as we used to,
because one of the proxied operations is also Pwait().  The proxying is just as
simple at this end, litle more than a function call and a couple of lines of
marshalling.  We set dpr_proxy_rq to NULL afterwards as a signal that the proxy
processing is complete (since it is not safe to use any condition variable
without wrapping it in a loop that checks some other variable, to guard against
spurious wakeups).

Unfortunately, that's not quite enough.  There are locking considerations,
because Ptrace() and Puntrace() have to be neatly balanced (even if proxied),
and the control thread spends time in libproc too.  If we didn't prevent it, the
main thread could do a Ptrace() and then the control thread could Puntrace() it,
leading to disaster.  So we have to take out a lock around Ptrace()/Puntrace()
using the recently-added Ptrace lock hook.

Unfortuanately we use this lock for other things, too, so we suddenly need it to
be a recursive mutex so we can take it multiple times without agonizing over the
global state of other lock holders. Still more unfortunately, we can't use an
real recursive mutex because you cannot use those with condition variables.  So
we convert dpr_lock into a hand-rolled recursive mutex using the dt_proc_t's
dpr_lock_holder and dpr_lock_count fields.  We use the ptrace hook mechanism to
automatically take that lock out across Ptrace()/Puntrace() calls, and arrange
to use the same new locking functions in most other places too.  (The conversion
is not complete: a few places cannot use it, or it simply didn't seem worth
making them use it because we know very well that these functions can never
be called while the same thread owns the mutex, even indirectly.)

This leads to some extra serialization. Among other things, it is impossible for
the main thread to interfere with most of the setup work of the control thread,
both because it holds the lock and because it hasn't got to a state where it can
process proxy requests yet.  The control thread drops the lock while poll()ing,
so that when it is ready to handle proxy requests, the lock is released so other
threads can call into libporc and send it proxy requests.

Finally, Solaris DTrace has the concept of an 'unmonitored process', which is a
process that was grabbed only briefly and doesn't have a monitor thread.  We
don't have that concept, so instead we add a new flag to the creation and
grabbing functions, DT_PROC_WAITING, which indicates that those functions should
automatically do a dt_proc_continue() and wait for the control thread to reach
its rendezvous event, following which it is ready to handle incoming libproc
requests properly.

On the testsuite front, we add a results preprocessor for the ustack test,
which discards lines with no symbols and lines from ld.so from the stack trace,
in the hope that the rest is relatively invariant across libc versions.

+++ libproc/libproc.h
@@ -121,7 +121,7 @@ extern	void	Puntrace(struct ps_prochandle *, int state);
 extern	void	Pclose(struct ps_prochandle *);
 
 extern	int	Pmemfd(struct ps_prochandle *);
-extern	int	Pwait(struct ps_prochandle *, boolean_t block);
+extern	long	Pwait(struct ps_prochandle *, boolean_t block);
 extern	int	Pstate(struct ps_prochandle *);
 extern	ssize_t	Pread(struct ps_prochandle *, void *, size_t, uintptr_t);
 extern	ssize_t Pread_string(struct ps_prochandle *, char *, size_t, uintptr_t);
@@ -160,7 +160,7 @@ extern	void	Pset_pwait_wrapper(struct ps_prochandle *P, pwait_fun *wrapper);
  * function should end up calling (somehow, from some thread or other).  Safe to
  * call only from the thread that did Pgrab() or Pcreate().
  */
-extern  int	Pwait_internal(struct ps_prochandle *P, boolean_t block);
+extern  long	Pwait_internal(struct ps_prochandle *P, boolean_t block);
 
 /*
  * Register a function to be called around the outermost layer of Ptrace()/
